update user profile
Рефакторинг кода и компонентов
