+++ src/components/personalAccount/profileUser/ProfileUser.jsx
@@ -1,33 +1,64 @@
-import { useEffect, useState } from "react";
+import "../profileUser/ProfileUser.scss";
 import { useUser } from "../../../context/UserContext";
-import ProfileClientInfo from "../profileClientInfo/ProfileClientInfo";
-import ProfileClinicInfo from "../profileClinicInfo/ProfileClinicInfo";
-import ProfileDoctorInfo from "../profileDoctorInfo/ProfileDoctorInfo";
-import AlertModal from "../../UI/modals/alertModal/AlertModal";
+import ProfileInfo from "../profileInfo/ProfileInfo";
 
 const ProfileUser = () => {
   const { userData } = useUser();
 
-  let profileInfoComponent;
+  const handleDeleteAccount = () => {
+    console.log("Account deleting");
+  };
+
+  let fieldsDetails = [];
+  let fieldsFooter = [];
+
   if (userData) {
     switch (userData?.user?.type) {
       case "client":
-        profileInfoComponent = <ProfileClientInfo userData={userData} />;
+        fieldsDetails = [
+          { label: "Пол", value: "gender" },
+          { label: "Дата рождения", value: "date_born" },
+          { label: "Почта", value: "email" },
+          { label: "Адрес", value: "address" },
+        ];
         break;
       case "clinic":
-        profileInfoComponent = <ProfileClinicInfo userData={userData} />;
+        fieldsDetails = [
+          { label: "Официальный сайт", value: "site" },
+          { label: "Почта", value: "email" },
+          { label: "Адрес", value: "address" },
+        ];
+        fieldsFooter = [{ label: "Лицензии и сертификаты", value: "licenses" }];
         break;
       case "surgeon":
-        profileInfoComponent = <ProfileDoctorInfo userData={userData} />;
+        fieldsDetails = [
+          { label: "Пол", value: "gender" },
+          { label: "Дата рождения", value: "date_born" },
+          { label: "Адрес", value: "address" },
+          { label: "Официальный сайт", value: "site" },
+        ];
+        fieldsFooter = [
+          { label: "Лицензии и сертификаты", value: "licenses" },
+          { label: "Образование", value: "education" },
+          { label: "Повышение квалификации", value: "qualification" },
+          { label: "Опыт работы", value: "experience" },
+        ];
         break;
       default:
-        profileInfoComponent = "Неизвестный тип пользователя";
+        return "Неизвестный тип пользователя";
     }
   } else {
-    profileInfoComponent = "Данные о пользователе не получены. Сервер в спячке";
+    return "Данные о пользователе не получены. Сервер в спячке";
   }
 
-  return <>{profileInfoComponent}</>;
+  return (
+    <ProfileInfo
+      userData={userData}
+      handleDeleteAccount={handleDeleteAccount}
+      fieldsDetails={fieldsDetails}
+      fieldsFooter={fieldsFooter}
+    />
+  );
 };
 
 export default ProfileUser;
